extract _RetryQueue class from matrix/transport.py into utils.py, remove transport.py
