+++ transport/matrix/node.py
@@ -1,9 +1,10 @@
 import json
 from collections import defaultdict
-from typing import Optional, Iterable, List, Dict, Tuple, Callable, Union, cast, Any, Set
+from typing import Optional, Iterable, List, Dict, Tuple, Callable, Union, cast, Any, Set, NewType
 from urllib.parse import urlparse
 
 import gevent
+import structlog
 from eth_utils import to_normalized_address, is_binary_address, to_checksum_address, to_canonical_address
 from gevent._semaphore import Semaphore
 from gevent.event import Event
@@ -25,13 +26,16 @@ from raiden.utils import Address, pex
 from raiden.utils.runnable import Runnable
 from raiden.utils.typing import ChainID, AddressHex
 from transport.matrix.client import Room, GMatrixClient
-from transport.matrix.transport import log, _RetryQueue, _RoomID
+from transport.utils import _RetryQueue
 from transport.matrix.utils import get_available_servers_from_config, make_client, get_server_url, UserAddressManager, \
     login_or_register, make_room_alias, join_global_room, UserPresence, validate_userid_signature, JOIN_RETRIES, \
     AddressReachability, validate_and_parse_message, login_or_register_light_client
 from transport.message import Message as TransportMessage
 from transport.node import Node as TransportNode
 from transport.udp import utils as udp_utils
+
+_RoomID = NewType("_RoomID", str)
+log = structlog.get_logger(__name__)
 
 
 class MatrixNode(TransportNode, Runnable):
